navigationbar, picture registration page and fix illegal  api storing problem
