+++ foodapp/src/App.js
@@ -1,64 +1,26 @@
-import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom'
-import SearchPage from './components/SearchPage'
-import HomePage from './components/HomePage'
-import RecipeViewPage from './components/RecipeViewPage'
-import UserViewPage from './components/UserViewPage'
-import LoginPage from './components/loginPage'
-import { AppBar, Toolbar, Button, Container } from '@mui/material'
-import RegistrationForm from './components/RegistrationPage'
-import UsersPage from './components/UsersPage'
-import UserProfile from './components/UserProfilePage'
-import CreateRecipePage from './components/createRecipe'
-import UserSearchPage from './components/UserRecipeSearch'
-//import RequireLogin from './components/RequireLogin'
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
+import SearchPage from './pages/SearchPage'
+import HomePage from './pages/HomePage'
+import RecipeViewPage from './pages/RecipeViewPage'
+import UserViewPage from './pages/UserViewPage'
+import LoginPage from './pages/loginPage'
+import { Container } from '@mui/material'
+import RegistrationForm from './pages/RegistrationPage'
+import UsersPage from './pages/UsersPage'
+import UserProfile from './pages/UserProfilePage'
+import CreateRecipePage from './pages/createRecipe'
+import UserSearchPage from './pages/UserRecipeSearch'
 import PersistedLogin from './components/PersistedLogin'
-import { selectCurrentUser } from './services/loginSlice'
-import { useSelector } from 'react-redux'
 import ErrorLayout from './components/ErrorLayout'
+import NavigationBar from './components/NavigationBar'
+
+
 
 const App = () => {
-  const user = useSelector(selectCurrentUser)
   return (
     <Router>
       <Container>
-        <AppBar position="static">
-          <Toolbar>
-            <Button color="inherit" component={Link} to="/">
-              <img src="/vectors/logo.svg" alt="Logo" style={{ height: '50px' }}/>
-            </Button>
-            <Button color="inherit" component={Link} to="/search">
-              search
-            </Button>
-            <Button color="inherit" component={Link} to="/users">
-              Discover users
-            </Button>
-            <Button color="inherit" component={Link} to="/userRecipesearch">
-              recipes by other users
-            </Button>
-            {user && (
-              <Button color="inherit" component={Link} to="/profile">
-                  My Profile
-              </Button>
-            )}
-            <Button color="inherit" component={Link} to="/users">
-                Discover users
-            </Button>
-            {user && (
-              <Button color="inherit" component={Link} to="/profile">
-                    My Profile
-              </Button>
-            )}
-            {user && (
-              <Button color="inherit" component={Link} to="/createrecipe">
-                    Create recipe
-              </Button>
-            )}
-            <Button color="inherit" component={Link} to="/login">
-                login
-            </Button>
-          </Toolbar>
-        </AppBar>
-
+        <NavigationBar/>
         <ErrorLayout>
           <Routes>
             <Route element={ <PersistedLogin />}>
